Revert "Add new bulk copy timeout test, add jmockit library"
This reverts commit 1c1b0820d8b4e65d9e6a51c34e8c03464fa5f7bf.
