+++ pom.xml
@@ -38,7 +38,6 @@
 	</scm>
 
 	<properties>
-        <jmockit.version>1.42</jmockit.version>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<junit.platform.version>1.2.0</junit.platform.version>
 		<junit.jupiter.version>5.2.0</junit.jupiter.version>
@@ -120,12 +119,6 @@
 			<version>2.5.0</version>
 			<scope>test</scope>
 		</dependency>
-        <dependency>
-            <groupId>org.jmockit</groupId>
-            <artifactId>jmockit</artifactId>
-            <version>${jmockit.version}</version>
-            <scope>test</scope>
-        </dependency>
 	</dependencies>
 
 	<profiles>
@@ -308,9 +301,6 @@
 				<version>2.22.0</version>
 
 				<configuration>
-                    <argLine>
-                        -javaagent:${settings.localRepository}/org/jmockit/jmockit/${jmockit.version}/jmockit-${jmockit.version}.jar
-                    </argLine>
 					<properties>
 						<excludeTags>${skipTestTag}</excludeTags>
 					</properties>
